chore(gh action): valideer functionele specs action
publiceert een summary van de validatie in de workflow job summary
