+++ scripts/generate-step-summary.js
@@ -1,31 +1,14 @@
 const fs = require('fs');
 const path = require('path');
+const { processFile } = require('./process-cucumber-file');
 
-function processFile(inputPath, outputPath) {
-    const content = fs.readFileSync(inputPath, 'utf8');
-    const lines = content.split('\n');
-    const results = [];
-    let footerLines = [];
+const outputFile = path.join(__dirname, './../test-reports/cucumber-js/step-summary.txt');
+fs.writeFileSync(outputFile, '', 'utf8');
 
-    if (lines.length >= 4) {
-        footerLines = lines.slice(-4);
-    }
+const fileMap = new Map([
+    ["./../test-reports/cucumber-js/step-definitions/test-result-zonder-dependency-integratie-summary.txt", "docs (zonder integratie)"],
+]);
 
-    lines.forEach((line) => {
-        const match = line.match(/^\d+\) Scenario: .*# (.+:\d+)/);
-        if (match) {
-            results.push((results.length + 1) + ": " + match[1]);
-        }
-    });
-
-    let finalOutput = footerLines.join('\n') + '\n\n';
-    if (results.length > 0) {
-        finalOutput += `<details>\n<summary>Bestandsnamen en regelnummers</summary>\n\n`;
-        finalOutput += results.join('\n') + '\n';
-        finalOutput += `\n</details>\n`;
-    }
-
-    fs.writeFileSync(outputPath, finalOutput, 'utf8');
-}
-
-processFile(process.argv[2], process.argv[3]);
+fileMap.forEach((caption, filePath) => {
+    processFile(path.join(__dirname, filePath), outputFile, caption);
+});
